chore: migrate BGP multidut folder
